Introduce Index super class for Indexer and Searcher
